+++ composer.json
@@ -1,0 +1,18 @@
+{
+    "name": "stokato/taskforce",
+    "description": "TaskForce project",
+    "type": "project",
+    "license": "free",
+    "authors": [
+        {
+            "name": "stokato",
+            "email": "pavel@demjanov.ru"
+        }
+    ],
+    "autoload": {
+      "psr-4": {
+        "Taskforce\\": "src/"
+      }
+    },
+    "require": {}
+}
